Tweak split positions code to avoid any size bloat
Make the split point an unsigned int instead of a size_t so the
structure is the same size as before this change.  It's possible
with 64-bit termpos for the split point to not be representable,
but only when there are more than 4 billion postings for the same
term in a single document, which isn't realistically going to
happen in non-pathological data.  If we can't store the split
point, we fall back to inserting into the vector like we did
before the split positions optimisation.

+++ xapian-core/api/terminfo.h
@@ -42,8 +42,12 @@ class TermInfo {
      *  If positions.empty(), then split > 0 indicates that the term has been
      *  deleted (this allows us to delete terms without invalidating existing
      *  TermIterator objects).
+     *
+     *  Use type unsigned here to avoid bloating this structure.  More than
+     *  4 billion positions in one document is not sensible (and not possible
+     *  unless termpos is configured to be 64 bit).
      */
-    mutable size_t split = 0;
+    mutable unsigned split = 0;
 
     /** Positions at which the term occurs.
      *
